requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.dev.in
#
alabaster==0.7.12         # via sphinx
babel==2.8.0              # via sphinx
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
docutils==0.16            # via m2r, sphinx
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
jinja2==2.11.2            # via numpydoc, sphinx
m2r==0.2.1                # via -r requirements.dev.in
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via m2r
numpydoc==1.1.0           # via -r requirements.dev.in
packaging==20.4           # via sphinx
pip-tools==5.3.1          # via -r requirements.dev.in
pygments==2.7.2           # via sphinx
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via babel
requests==2.24.0          # via sphinx
six==1.15.0               # via packaging, pip-tools
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.5.0   # via -r requirements.dev.in
sphinx==3.2.1             # via -r requirements.dev.in, numpydoc, sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
urllib3==1.25.11          # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
